Move location of use mute user hooks
